+++ src/parse_bench/inference/providers/parse/docling_serve.py
@@ -1,4 +1,4 @@
-"""Provider for Docling serve via a custom HTTP endpoint."""
+"""Provider for Docling via the official docling-serve HTTP API."""
 
 import base64
 import os
@@ -19,14 +19,9 @@ from parse_bench.inference.providers.base import (
     ProviderRateLimitError,
     ProviderTransientError,
 )
+from parse_bench.inference.providers.parse._docling_common import _build_docling_layout_pages
 from parse_bench.inference.providers.registry import register_provider
-from parse_bench.layout_label_mapping import (
-    UnknownRawLayoutLabelError,
-    map_docling_raw_label_to_canonical,
-)
 from parse_bench.schemas.parse_output import (
-    LayoutItemIR,
-    LayoutSegmentIR,
     PageIR,
     ParseLayoutPageIR,
     ParseOutput,
@@ -39,233 +34,16 @@ from parse_bench.schemas.pipeline_io import (
 )
 from parse_bench.schemas.product import ProductType
 
-_DOCLING_EXCLUDED_LAYOUT_LABELS = frozenset(
-    {
-        "empty_value",
-        "field_heading",
-        "field_hint",
-        "field_item",
-        "field_key",
-        "field_region",
-        "field_value",
-        "marker",
-    }
-)
-_DOCLING_TABLE_LABELS = frozenset({"document_index", "table"})
-_DOCLING_IMAGE_LABELS = frozenset({"chart", "picture"})
 _MD_PAGE_BREAK_PLACEHOLDER = "<!-- page-break -->"
-
-
-def _normalize_docling_label(label: object) -> str | None:
-    if label is None:
-        return None
-    value = getattr(label, "value", label)
-    if not isinstance(value, str):
-        return None
-    return value.strip().lower()
-
-
-def _should_include_docling_label(raw_label: str) -> bool:
-    if raw_label in _DOCLING_EXCLUDED_LAYOUT_LABELS:
-        return False
-    try:
-        map_docling_raw_label_to_canonical(raw_label)
-    except UnknownRawLayoutLabelError:
-        return False
-    return True
-
-
-def _docling_item_type(raw_label: str) -> str:
-    if raw_label in _DOCLING_TABLE_LABELS:
-        return "table"
-    if raw_label in _DOCLING_IMAGE_LABELS:
-        return "image"
-    return "text"
-
-
-def _extract_docling_item_value(item: Any, doc: DoclingDocument, raw_label: str) -> str:
-    item_type = _docling_item_type(raw_label)
-    if item_type == "image":
-        return ""
-
-    if item_type == "table" and hasattr(item, "export_to_html"):
-        try:
-            html = item.export_to_html(doc=doc, add_caption=True)
-            if isinstance(html, str):
-                return html
-        except Exception:
-            pass
-
-    text = getattr(item, "text", None)
-    if isinstance(text, str):
-        return text
-
-    if hasattr(item, "export_to_markdown"):
-        try:
-            markdown = item.export_to_markdown()
-            if isinstance(markdown, str):
-                return markdown
-        except Exception:
-            pass
-
-    return ""
-
-
-def _normalize_docling_charspan(
-    charspan: object,
-    *,
-    text_length: int,
-    include_span: bool,
-) -> tuple[int | None, int | None]:
-    if not include_span or not isinstance(charspan, (list, tuple)) or len(charspan) != 2:
-        return (None, None)
-
-    start_raw, end_raw = charspan
-    if not isinstance(start_raw, int) or not isinstance(end_raw, int):
-        return (None, None)
-
-    start = max(0, min(start_raw, text_length))
-    end_exclusive = max(start, min(end_raw, text_length))
-    if end_exclusive <= start:
-        return (None, None)
-
-    # Docling charspan behaves like a Python slice [start, end).
-    return (start, end_exclusive - 1)
-
-
-def _build_docling_segment(
-    *,
-    prov: Any,
-    raw_label: str,
-    page_width: float,
-    page_height: float,
-    include_span: bool,
-    text_length: int,
-) -> LayoutSegmentIR | None:
-    bbox = getattr(prov, "bbox", None)
-    if bbox is None or page_width <= 0 or page_height <= 0:
-        return None
-
-    bbox_top_left = bbox.to_top_left_origin(page_height=page_height)
-    width = bbox_top_left.r - bbox_top_left.l
-    height = bbox_top_left.b - bbox_top_left.t
-    if width <= 0 or height <= 0:
-        return None
-
-    start_index, end_index = _normalize_docling_charspan(
-        getattr(prov, "charspan", None),
-        text_length=text_length,
-        include_span=include_span,
-    )
-
-    return LayoutSegmentIR(
-        x=bbox_top_left.l / page_width,
-        y=bbox_top_left.t / page_height,
-        w=width / page_width,
-        h=height / page_height,
-        confidence=1.0,
-        label=raw_label,
-        start_index=start_index,
-        end_index=end_index,
-    )
-
-
-def _merge_segments(segments: list[LayoutSegmentIR]) -> LayoutSegmentIR | None:
-    if not segments:
-        return None
-
-    x1 = min(segment.x for segment in segments)
-    y1 = min(segment.y for segment in segments)
-    x2 = max(segment.x + segment.w for segment in segments)
-    y2 = max(segment.y + segment.h for segment in segments)
-    return LayoutSegmentIR(
-        x=x1,
-        y=y1,
-        w=x2 - x1,
-        h=y2 - y1,
-        confidence=1.0,
-        label=segments[0].label,
-    )
-
-
-def _build_docling_layout_pages(
-    *,
-    doc: DoclingDocument,
-    raw_pages: list[dict[str, Any]],
-) -> list[ParseLayoutPageIR]:
-    page_markdown_by_number: dict[int, str] = {}
-    for page_data in raw_pages:
-        page_number = page_data.get("page")
-        if isinstance(page_number, int) and page_number > 0:
-            page_markdown_by_number[page_number] = str(page_data.get("markdown", ""))
-
-    layout_pages: list[ParseLayoutPageIR] = []
-    for page_number in sorted(doc.pages.keys()):
-        page = doc.pages[page_number]
-        page_width = float(page.size.width)
-        page_height = float(page.size.height)
-        items: list[LayoutItemIR] = []
-
-        for item, _level in doc.iterate_items(page_no=page_number):
-            raw_label = _normalize_docling_label(getattr(item, "label", None))
-            if raw_label is None or not _should_include_docling_label(raw_label):
-                continue
-
-            item_type = _docling_item_type(raw_label)
-            item_value = _extract_docling_item_value(item, doc, raw_label)
-            include_span = item_type == "text"
-
-            page_provs = [
-                prov for prov in getattr(item, "prov", []) or [] if getattr(prov, "page_no", None) == page_number
-            ]
-            segments = [
-                segment
-                for prov in page_provs
-                if (
-                    segment := _build_docling_segment(
-                        prov=prov,
-                        raw_label=raw_label,
-                        page_width=page_width,
-                        page_height=page_height,
-                        include_span=include_span,
-                        text_length=len(item_value),
-                    )
-                )
-                is not None
-            ]
-            if not segments:
-                continue
-
-            merged_bbox = _merge_segments(segments)
-            items.append(
-                LayoutItemIR(
-                    type=item_type,
-                    value=item_value,
-                    bbox=merged_bbox,
-                    layout_segments=segments,
-                )
-            )
-
-        layout_pages.append(
-            ParseLayoutPageIR(
-                page_number=page_number,
-                width=page_width,
-                height=page_height,
-                md=page_markdown_by_number.get(page_number, ""),
-                items=items,
-            )
-        )
-
-    return layout_pages
 
 
 @register_provider("docling_serve")
 class DoclingServeProvider(Provider):
     """
-    Provider for Docling PDF parsing via a custom HTTP endpoint.
+    Provider for Docling PDF parsing via the official docling-serve HTTP API.
 
-    This provider sends PDFs to a Docling endpoint and returns markdown
-    with tables formatted as HTML.
+    This provider sends PDFs to the docling-serve HTTP API endpoint and returns markdown
+    with tables formatted as HTML. It was tested with docling-serve v1.17.0.
     """
 
     def __init__(
@@ -274,30 +52,22 @@ class DoclingServeProvider(Provider):
         base_config: dict[str, Any] | None = None,
     ):
         """
-        Initialize the Docling serve provider.
+        Initialize the Docling Serve provider.
 
         Args:
             provider_name: Name of the provider
             base_config: Optional configuration with:
                 - `api_key`: Optional bearer token for the endpoint
-                - `hf_token`: Deprecated alias for `api_key`
                 - `endpoint_url`: Endpoint URL (required)
                 - `timeout`: Request timeout in seconds (default: 120)
         """
         super().__init__(provider_name, base_config)
 
-        # Optional bearer token; keep `hf_token` / `HF_TOKEN` as a deprecated
-        # fallback for backwards compatibility with the previous HF deployment.
-        self._api_key = (
-            self.base_config.get("api_key")
-            or self.base_config.get("hf_token")
-            or os.getenv("DOCLING_SERVE_API_KEY")
-            or os.getenv("HF_TOKEN")
-            or ""
-        )
+        self._api_key = self.base_config.get("api_key") or os.getenv("DOCLING_SERVE_API_KEY") or ""
 
         # Get endpoint URL (from config or env var)
         self._endpoint_url = self.base_config.get("endpoint_url") or os.getenv("DOCLING_SERVE_ENDPOINT_URL")
+        self._endpoint_url = self._endpoint_url.rstrip("/")
         if not self._endpoint_url:
             raise ProviderConfigError(
                 "Docling Serve endpoint URL is required. "
@@ -314,6 +84,7 @@ class DoclingServeProvider(Provider):
 
         Args:
             pdf_bytes: Raw PDF file bytes
+            filename: Name of the PDF file
 
         Returns:
             Raw JSON response from endpoint
@@ -374,7 +145,12 @@ class DoclingServeProvider(Provider):
             raise ProviderTransientError(f"Connection error: {e}") from e
         except requests.exceptions.HTTPError as e:
             status_code = e.response.status_code if e.response else None
-            if status_code == 429:
+            if status_code == 422:
+                raise ProviderPermanentError(
+                    "Docling Serve returned 422. Ensure docling-serve >= 1.0 "
+                    f"(older versions expect 'file_sources' instead of 'sources'): {e}"
+                ) from e
+            elif status_code == 429:
                 raise ProviderRateLimitError(f"Rate limit exceeded: {e}") from e
             elif status_code and 500 <= status_code < 600:
                 raise ProviderTransientError(f"Server error ({status_code}): {e}") from e
@@ -403,7 +179,7 @@ class DoclingServeProvider(Provider):
         """
         if request.product_type != ProductType.PARSE:
             raise ProviderPermanentError(
-                f"DoclingParseProvider only supports PARSE product type, got {request.product_type}"
+                f"DoclingServeProvider only supports PARSE product type, got {request.product_type}"
             )
 
         started_at = datetime.now()
@@ -464,7 +240,7 @@ class DoclingServeProvider(Provider):
         #   }
         # }
         full_markdown = ""
-        raw_docling_document = raw_result.raw_output.get("document", {}).get("json_content", "")
+        raw_docling_document = raw_result.raw_output.get("document", {}).get("json_content")
         pages: list[PageIR] = []
 
         layout_pages: list[ParseLayoutPageIR] = []
@@ -480,15 +256,16 @@ class DoclingServeProvider(Provider):
             full_markdown = doc_serializer.serialize(
                 page_break_placeholder=_MD_PAGE_BREAK_PLACEHOLDER, image_mode=ImageRefMode.PLACEHOLDER
             ).text
-            raw_pages = full_markdown.split(_MD_PAGE_BREAK_PLACEHOLDER)
+            raw_pages_md = full_markdown.split(_MD_PAGE_BREAK_PLACEHOLDER)
+            raw_pages_dicts = []
 
-            # Convert to PageIR list (0-indexed)
-            for page_index, markdown in enumerate(raw_pages):
+            for page_index, markdown in enumerate(raw_pages_md):
                 pages.append(PageIR(page_index=page_index, markdown=markdown))
+                raw_pages_dicts.append({"page": page_index + 1, "markdown": markdown})
 
             layout_pages = _build_docling_layout_pages(
                 doc=docling_document,
-                raw_pages=[page for page in raw_pages if isinstance(page, dict)],
+                raw_pages=raw_pages_dicts,
             )
 
         output = ParseOutput(
